+++ test/Confluent.Kafka.IntegrationTests/Tests/Consumer_Assignment.cs
@@ -36,7 +36,7 @@ namespace Confluent.Kafka.IntegrationTests
             LogToFile("start Consumer_Assignment");
 
             int N = 2;
-            var firstProduced = Util.ProduceMessages(bootstrapServers, singlePartitionTopic, 1, N);
+            var firstProduced = Util.ProduceNullStringMessages(bootstrapServers, singlePartitionTopic, 1, N);
 
             var consumerConfig = new ConsumerConfig
             {